+++ components/yay_com/actions/get-documents/get-documents.mjs
@@ -1,18 +1,26 @@
-import yay_com from "../../app/yay_com.app.mjs";
+import yay_com from "../../yay_com.app.mjs";
 
 export default {
   key: "yay_com-get-documents",
   name: "Get Documents",
-  description: "Retrieves a list of documents. [See documentation](https://www.yay.com/voip/api-docs/account/document/)",
+  description:
+    "Retrieves a list of documents. [See documentation](https://www.yay.com/voip/api-docs/account/document/)",
   version: "0.0.1",
   type: "action",
   props: {
     yay_com,
   },
   async run({ $ }) {
-    const response = await this.yay_com.listDocuments($);
-    const documents = response.data || [];
-    $.export("$summary", `Successfully retrieved ${documents.length} document(s)`);
+    const response = await this.yay_com.listDocuments({
+      $,
+    });
+    const { length } = response;
+    $.export(
+      "$summary",
+      `Successfully retrieved ${length} document${length === 1
+        ? ""
+        : "s"}`,
+    );
     return response;
   },
 };
